Graph threading graph-> macro now threads through any form, not just ones in the graph namespaces. This is because it's entirely reasonable to expect users will want to define their own accessors and have them plug into the graph-> macro, which is not possible with the namespace-hackery.
However, to allow users to *not* have a piece of code auto-threaded, there is now the graph-| "macro" (not actually a macro, but pretends like it is, ala unquote). When used inside a graph-> statement, its inside is not parsed by the threading insertion routines, and is instead converted into an implicit do statement.